requirements.txt
tsic
psutil
paho-mqtt
pigpio
Kivy
kivy-garden.graph
pyserial

# install this manually AFTER other deps:
#https://github.com/kivy-garden/graph/archive/master.zip

# patch kivy/logger.py and comment out change of the kivy root logger like this:
#logging.root = Logger
